+++ typings/mysql/lib/protocol/packets/index.d.ts
@@ -2,6 +2,7 @@ import { OkPacket } from './OkPacket.js';
 import { RowDataPacket } from './RowDataPacket.js';
 import { FieldPacket } from './FieldPacket.js';
 import { Field } from './Field.js';
+import { ProcedureCallPacket } from './ProcedurePacket.js';
 import { ResultSetHeader } from './ResultSetHeader.js';
 import { OkPacketParams } from './params/OkPacketParams.js';
 import { ErrorPacketParams } from './params/ErrorPacketParams.js';
@@ -11,6 +12,7 @@ export {
   RowDataPacket,
   FieldPacket,
   Field,
+  ProcedureCallPacket,
   ResultSetHeader,
   OkPacketParams,
   ErrorPacketParams,